refactory to reduce responsibility of HBaseRelation
